Remove ValueError being raised in get_cover of volume.
If an invalid product id is passed as a paremeter get_cover now returns None
instead of raising a value error. Friedbert will then render a reference with
a fallback cover not nothing. In addition to this change a new validation
rule is added to the cp-validation rules, which raises an error if a wrong
volume is referenced in an article and warns the user in the UI.